+алгоритм получения адреса по локации

+++ WebApi/Dto/Request/GetAddressByLocationDto.cs
@@ -9,7 +9,7 @@ public class GetAddressByLocationDto
     [Required(ErrorMessage = "Wrong longitude")]
     public double Long { get; set; }
     
-    public double? Radius { get; set; }
+    public float? Radius { get; set; }
     public int? Page { get; set; }
     public int? PageSize { get; set; }
 }
